# Add option to use in-memory FS for pebble