--- a/aggregators/aggregator_test.go
+++ b/aggregators/aggregator_test.go
@@ -1199,7 +1199,7 @@ func newTestAggregator(tb testing.TB) *Aggregator {
 
 func flushTestAggregator(tb testing.TB, agg *Aggregator) {
 	if agg.batch != nil {
-		if err := agg.batch.Commit(pebble.Sync); err != nil {
+		if err := agg.batch.Commit(agg.cfg.WriteOpts); err != nil {
 			tb.Fatal(err)
 		}
 		if err := agg.batch.Close(); err != nil {