--- a/aggregators/config.go
+++ b/aggregators/config.go
@@ -11,7 +11,6 @@ import (
 	"sort"
 	"time"
 
-	"github.com/cockroachdb/pebble"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
@@ -45,7 +44,6 @@ type Config struct {
 	HarvestDelay           time.Duration
 	CombinedMetricsIDToKVs func([16]byte) []attribute.KeyValue
 	InMemory               bool
-	WriteOpts              *pebble.WriteOptions
 
 	Meter  metric.Meter
 	Tracer trace.Tracer
@@ -188,7 +186,6 @@ func defaultCfg() Config {
 		Tracer:                 otel.Tracer(instrumentationName),
 		CombinedMetricsIDToKVs: func(_ [16]byte) []attribute.KeyValue { return nil },
 		Logger:                 zap.Must(zap.NewDevelopment()),
-		WriteOpts:              pebble.Sync,
 	}
 }
 
@@ -227,10 +224,6 @@ func validateCfg(cfg Config) error {
 	}
 	if highest > 18*time.Hour {
 		return errors.New("aggregation interval greater than 18 hours is not supported")
-	}
-
-	if cfg.WriteOpts != pebble.Sync && cfg.WriteOpts != pebble.NoSync {
-		return errors.New("write opts should be pebble.Sync or pebble.NoSync")
 	}
 	return nil
 }