Implemented as a boolean config to avoid exposing pebble internals / config to users.