@carsonip 2023-07-13
> Did you run any tests with the memory vfs?

I manually ran TestAggregateAndHarvest and it works ok. I ran the benchmarks with memory vfs for inmemoryfs-after.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/inmemoryfs-before │     bench-out/inmemoryfs-after      │
                            │           sec/op            │    sec/op     vs base               │
AggregateCombinedMetrics-16                   3.434µ ± 4%    3.213µ ± 4%   -6.41% (p=0.004 n=6)
AggregateBatchSerial-16                       13.95µ ± 6%    13.22µ ± 8%        ~ (p=0.180 n=6)
AggregateBatchParallel-16                     14.26µ ± 2%    15.20µ ± 9%   +6.64% (p=0.026 n=6)
CombinedMetricsEncoding-16                    5.218µ ± 4%    5.177µ ± 2%        ~ (p=0.310 n=6)
CombinedMetricsDecoding-16                    4.327µ ± 1%    4.961µ ± 4%  +14.64% (p=0.002 n=6)
CombinedMetricsToBatch-16                     431.9µ ± 6%    425.2µ ± 2%        ~ (p=0.065 n=6)
EventToCombinedMetrics-16                     966.7n ± 5%   1130.5n ± 9%  +16.94% (p=0.002 n=6)
geomean                                       9.390µ         9.682µ        +3.11%

                            │ bench-out/inmemoryfs-before │      bench-out/inmemoryfs-after       │
                            │            B/op             │     B/op      vs base                 │
AggregateCombinedMetrics-16                3.443Ki ± 1%     3.878Ki ± 1%  +12.63% (p=0.002 n=6)
AggregateBatchSerial-16                    20.15Ki ± 0%     22.29Ki ± 1%  +10.62% (p=0.002 n=6)
AggregateBatchParallel-16                  20.10Ki ± 0%     22.60Ki ± 1%  +12.44% (p=0.002 n=6)
CombinedMetricsEncoding-16                   0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=6) ¹
CombinedMetricsDecoding-16                 10.01Ki ± 0%     10.01Ki ± 0%        ~ (p=0.182 n=6)
geomean                                                 ²                  +6.97%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ bench-out/inmemoryfs-before │     bench-out/inmemoryfs-after     │
                            │          allocs/op          │ allocs/op   vs base                │
AggregateCombinedMetrics-16                  27.00 ± 4%     27.00 ± 4%       ~ (p=1.000 n=6)
AggregateBatchSerial-16                      94.00 ± 0%     94.00 ± 0%       ~ (p=1.000 n=6) ¹
AggregateBatchParallel-16                    94.00 ± 0%     94.00 ± 1%       ~ (p=1.000 n=6)
CombinedMetricsEncoding-16                   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=6) ¹
CombinedMetricsDecoding-16                   40.00 ± 0%     40.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                 ²               +0.00%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

No idea why it affects EventToCombinedMetrics as it is not related to pebble at all.