@marclop 2023-07-14
Am I reading the results right? It appears to be slower with the in-memory VFS