@carsonip 2023-07-19
> Am I reading the results right? It appears to be slower with the in-memory VFS

The weird thing is that the performance drop cannot be observed when the affected benchmarks are run separately. It also doesn't make sense for them to have any performance hit because of in-memory VFS because BenchmarkCombinedMetricsDecoding does not touch anything about pebble.

I suspect it is gc (because the vfs requires memory allocations) causing noise to subsequent benchmarks.