@carsonip 2023-07-24
With NewMem and DisableWAL:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/new-mem-and-disable-wal-2-before │ bench-out/new-mem-and-disable-wal-2-after │
                            │                   sec/op                   │       sec/op        vs base               │
AggregateCombinedMetrics-16                                  4.006µ ± 2%         3.901µ ± 25%       ~ (p=0.190 n=10)
AggregateBatchSerial-16                                     10.561µ ± 5%         9.969µ ±  3%  -5.61% (p=0.001 n=10)
AggregateBatchParallel-16                                    10.81µ ± 6%         10.55µ ±  6%       ~ (p=0.165 n=10)
CombinedMetricsEncoding-16                                   4.925µ ± 2%         5.061µ ±  4%  +2.77% (p=0.011 n=10)
CombinedMetricsDecoding-16                                   4.394µ ± 2%         4.563µ ±  8%  +3.83% (p=0.001 n=10)
CombinedMetricsToBatch-16                                    411.5µ ± 1%         418.9µ ±  1%  +1.80% (p=0.001 n=10)
EventToCombinedMetrics-16                                    837.3n ± 5%         877.9n ±  3%  +4.85% (p=0.004 n=10)
geomean                                                      8.576µ              8.602µ        +0.30%

                            │ bench-out/new-mem-and-disable-wal-2-before │ bench-out/new-mem-and-disable-wal-2-after │
                            │                    B/op                    │       B/op        vs base                 │
AggregateCombinedMetrics-16                               4.857Ki ± 0%         4.885Ki ± 2%       ~ (p=0.171 n=10)
AggregateBatchSerial-16                                   11.34Ki ± 0%         11.31Ki ± 0%       ~ (p=0.065 n=10)
AggregateBatchParallel-16                                 11.29Ki ± 1%         11.34Ki ± 1%       ~ (p=0.353 n=10)
CombinedMetricsEncoding-16                                  0.000 ± 0%           0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                                10.01Ki ± 0%         10.01Ki ± 0%       ~ (p=0.885 n=10)
CombinedMetricsToBatch-16                                 37.66Ki ± 0%         37.66Ki ± 0%       ~ (p=1.000 n=10) ¹
EventToCombinedMetrics-16                                   16.00 ± 0%           16.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                ²                     +0.10%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ bench-out/new-mem-and-disable-wal-2-before │ bench-out/new-mem-and-disable-wal-2-after │
                            │                 allocs/op                  │    allocs/op      vs base                 │
AggregateCombinedMetrics-16                                 39.00 ± 0%           39.00 ± 3%       ~ (p=0.635 n=10)
AggregateBatchSerial-16                                     67.00 ± 0%           67.00 ± 1%       ~ (p=1.000 n=10)
AggregateBatchParallel-16                                   67.00 ± 0%           67.00 ± 0%       ~ (p=1.000 n=10)
CombinedMetricsEncoding-16                                  0.000 ± 0%           0.000 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                                  40.00 ± 0%           40.00 ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsToBatch-16                                   308.0 ± 0%           308.0 ± 0%       ~ (p=1.000 n=10) ¹
EventToCombinedMetrics-16                                   2.000 ± 0%           2.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                ²                     +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```